ENH: Improve FindLibUsb1 and make default on WIN32 too
1) Improve Findlibusb-1.0 :
- search for libusb.h path not libusb-1.0/libusb.h (PATH_SUFFIXES helps).
- search for usb-1.0 and libusb-1.0 as library names, the later being
used on windows.
- use find_package_handle_standard_args to ensure variables are found.
This results in a change of the camke variables to LIBUSB1_* instead
of LIBUSB_1_*.

2) Make Libusb1-0 default on Windows instead of LIBFTD2XX. LibUsb1 can
be easily downloaded from sourceforge and wraps the include file, libs
for different platforms and generators, static and shared :
https://sourceforge.net/projects/libusb/files/libusb-1.0/

For ease of use it would possible to download, archive, and install
files from libusb-1.0 easily using CMake ExternalProject_Add() macro.